pyproject.toml
[tool.poetry]
name = "landscape-analysis"
version = "0.0.1"
description = "Enabling landscape analysis to better understand existing efforts or solutions."
authors = ["Way Science community"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pandas = ">1.1,<1.4.0 || >1.4.0,<2.1"
numpy = ">=1.16.0,<1.26"
matplotlib = ">=3.2,<=3.7.3"
plotly = "^5.17.0"
duckdb = "^0.9.0"
requests = "^2.31.0"
pygithub = "^2.1.1"
python-box = "^7.1.1"
pytz = "^2023.3.post1"
poetry = "^1.6.1"
poethepoet = "^0.24.1"
pyarrow = "^13.0.0"
pyppeteer = "^1.0.2"
nest-asyncio = "^1.5.8"
htmlmin = "^0.1.12"
ydata-profiling = "^4.6.0"
ipywidgets = "^8.1.1"
itables = "^1.6.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.9.1"
jupyterlab-code-formatter = "^2.2.1"
jupytext = "^1.15.2"
jupyterlab = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# defines a workflow for cytomining ecosystem software landscape analysis
[tool.poe.tasks.cytomining_ecosystem]
# sets a cwd relative for the shell script below
cwd = "notebooks/cytomining-ecosystem"
# defines workflow through sequence of python scripts converted from jupyter notebooks
shell = """
  python seek-landscape-entries.py &&
  python project-github-metrics.py &&
  python visualize-landscape.py
"""

[tool.jupytext]
formats = "ipynb,py:light"

[tool.vulture]
min_confidence = 80
paths = ["notebooks"]
sort_by_size = true
verbose = true

[tool.isort]
profile = "black"
